+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [3.1.0] - 2026-07-06
 
 ### Changed
  - results JSON simplified (no aggregations etc., just list of results)
@@ -147,6 +147,7 @@ Initial release to setup PyPI.
 
 Fully functional emfuzzer - see README.md for details
 
+[3.1.0]: https://github.com/ZBOSK-II/emtorch/compare/3.0.1...3.1.0
 [3.0.1]: https://github.com/ZBOSK-II/emtorch/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/ZBOSK-II/emtorch/compare/2.1.2...3.0.0
 [2.1.2]: https://github.com/ZBOSK-II/emtorch/compare/2.1.1...2.1.2